Revert "Edited Homework17"
