+++ src/test/java/LoginTests.java
@@ -6,7 +6,7 @@ import org.testng.annotations.Test;
 
 import java.time.Duration;
 
-public class LoginTests{
+public class LoginTests extends BaseTest {
     @Test
     public static void LoginEmptyEmailPasswordTest() {
